+++ common/src/main/java/net/bettercombat/api/WeaponAttributesHelper.java
@@ -20,6 +20,7 @@ import java.util.ArrayList;
 public class WeaponAttributesHelper {
     public static WeaponAttributes override(WeaponAttributes a, WeaponAttributes b) {
         var attackRange = b.attackRange() > 0 ? b.attackRange() : a.attackRange();
+        var rangeBonus = b.rangeBonus() > 0 ? b.rangeBonus() : a.rangeBonus();
         var pose = b.pose() != null ? b.pose() : a.pose();
         var off_hand_pose = b.offHandPose() != null ? b.offHandPose() : a.offHandPose();
         var isTwoHanded = b.two_handed() != null ? b.two_handed() : a.two_handed();
@@ -45,7 +46,7 @@ public class WeaponAttributesHelper {
             }
             attacks = overrideAttacks.toArray(new WeaponAttributes.Attack[0]);
         }
-        return new WeaponAttributes(attackRange, pose, off_hand_pose, isTwoHanded, category, attacks);
+        return new WeaponAttributes(attackRange, rangeBonus, pose, off_hand_pose, isTwoHanded, category, attacks);
     }
 
     public static void validate(WeaponAttributes attributes) throws Exception {