+++ common/src/main/java/net/bettercombat/config/ClientConfig.java
@@ -28,6 +28,8 @@ public class ClientConfig implements ConfigData {
     @ConfigEntry.Gui.Tooltip
     public boolean isTooltipAttackRangeEnabled = true;
     @ConfigEntry.Gui.Tooltip
+    public boolean isTooltipAttackRangeDynamic = false;
+    @ConfigEntry.Gui.Tooltip
     @ConfigEntry.BoundedDiscrete(min = 0, max = 100)
     public int weaponSwingSoundVolume = 100;
     @ConfigEntry.Gui.Tooltip