+++ common/src/main/java/net/bettercombat/logic/PlayerAttackHelper.java
@@ -5,6 +5,7 @@ import net.bettercombat.api.AttackHand;
 import net.bettercombat.api.ComboState;
 import net.bettercombat.api.WeaponAttributes;
 import net.bettercombat.utils.AttributeModifierHelper;
+import net.minecraft.entity.attribute.EntityAttributes;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.item.ItemStack;
 import net.minecraft.item.ShieldItem;
@@ -225,5 +226,37 @@ public class PlayerAttackHelper {
             offPose = "";
         }
         return new Pose(mainPose, offPose);
+    }
+
+    public static double getRange(PlayerEntity player) {
+        return getRange(player, player.getMainHandStack());
+    }
+
+    public static double getRange(PlayerEntity player, ItemStack stack) {
+        var attributes = WeaponRegistry.getAttributes(stack);
+        return getRange(player, attributes);
+    }
+
+    public static double getStaticRange(PlayerEntity player, ItemStack stack) {
+        var attributes = WeaponRegistry.getAttributes(stack);
+        return getRange(player, attributes, false);
+    }
+
+    public static double getRange(PlayerEntity player, WeaponAttributes attributes) {
+        return getRange(player, attributes, true);
+    }
+
+    public static double getRange(PlayerEntity player, WeaponAttributes attributes, boolean realValue) {
+        var range = realValue
+                ? player.getEntityInteractionRange()
+                : EntityAttributes.PLAYER_ENTITY_INTERACTION_RANGE.value().getDefaultValue();
+        if (attributes != null) {
+            // Absolute range (legacy)
+            if (attributes.attackRange() != 0) {
+                return attributes.attackRange();
+            }
+            range += attributes.rangeBonus();
+        }
+        return range;
     }
 }